+++ src/components/about-film.js
@@ -1,5 +1,4 @@
 import {MONTHS} from '../const.js';
-// import {createElement} from '../utils.js';
 import AbstractComponent from './abstract-component.js';
 
 const createGenresMarkup = (genres) => {
@@ -166,32 +165,6 @@ const createAboutFilmPopupTemplate = (film, comments) => {
     </section>`
   );
 };
-
-
-// export default class AboutFilmPopupComponent {
-//   constructor(film, comments) {
-//     this._film = film;
-//     this._comments = comments;
-//     this._element = null;
-//   }
-
-//   getTemplate() {
-//     return createAboutFilmPopupTemplate(this._film, this._comments);
-//   }
-
-//   getElement() {
-//     if (!this._element) {
-//       this._element = createElement(this.getTemplate());
-//     }
-
-//     return this._element;
-//   }
-
-//   removeElement() {
-//     this._element = null;
-//   }
-// }
-
 
 export default class AboutFilmPopup extends AbstractComponent {
   constructor(film, comments) {
