+++ src/components/list-film-cards.js
@@ -1,4 +1,3 @@
-// import {createElement} from '../utils.js';
 import AbstractComponent from './abstract-component.js';
 
 // функция возвращающая Список карточек с фильмами
@@ -20,28 +19,6 @@ const createListFilmCardsTemplate = () => {
     </section>`
   );
 };
-
-// export default class ListFilmCardsComponent {
-//   constructor() {
-//     this._element = null;
-//   }
-
-//   getTemplate() {
-//     return createListFilmCardsTemplate();
-//   }
-
-//   getElement() {
-//     if (!this._element) {
-//       this._element = createElement(this.getTemplate());
-//     }
-
-//     return this._element;
-//   }
-
-//   removeElement() {
-//     this._element = null;
-//   }
-// }
 
 export default class ListFilmCards extends AbstractComponent {
   getTemplate() {
