+++ src/components/main-navigation.js
@@ -1,11 +1,9 @@
-// import {createElement} from '../utils.js';
 import AbstractComponent from './abstract-component.js';
 
 const createFilterMarkup = (filter) => {
   const {name, count} = filter;
   return name === `All movies` ? `<a href="#all" class="main-navigation__item main-navigation__item--active">${name}</a>` : `<a href="#watchlist" class="main-navigation__item">${name} <span class="main-navigation__item-count">${count}</span></a>`;
 };
-
 
 // функция возвращающая Меню
 const createMainNavigationTemplate = (filters) => {
@@ -18,35 +16,10 @@ ${filterMarkup}
   );
 };
 
-// export default class MainNavigationComponent {
-//   constructor(filters) {
-//     this._filters = filters;
-//     this._element = null;
-//   }
-
-//   getTemplate() {
-//     return createMainNavigationTemplate(this._filters);
-//   }
-
-//   getElement() {
-//     if (!this._element) {
-//       this._element = createElement(this.getTemplate());
-//     }
-
-//     return this._element;
-//   }
-
-//   removeElement() {
-//     this._element = null;
-//   }
-// }
-
-
 export default class MainNavigation extends AbstractComponent {
   constructor(filters) {
     super();
     this._filters = filters;
-    //  this._element = null;
   }
 
   getTemplate() {
