+++ src/components/no-data_film.js
@@ -1,4 +1,3 @@
-// import {createElement} from '../utils.js';
 import AbstractComponent from './abstract-component.js';
 
 // функция возвращающая сообщение, что нет фильмов
@@ -11,28 +10,6 @@ const createNoDataFilmTemplate = () => {
   </section>`
   );
 };
-
-// export default class NoDataFilmComponent {
-//   constructor() {
-//     this._element = null;
-//   }
-
-//   getTemplate() {
-//     return createNoDataFilmTemplate();
-//   }
-
-//   getElement() {
-//     if (!this._element) {
-//       this._element = createElement(this.getTemplate());
-//     }
-
-//     return this._element;
-//   }
-
-//   removeElement() {
-//     this._element = null;
-//   }
-// }
 
 export default class NoDataFilm extends AbstractComponent {
   getTemplate() {
