+++ src/components/sort.js
@@ -1,4 +1,3 @@
-// import {createElement} from '../utils.js';
 import AbstractComponent from './abstract-component.js';
 
 const createSortFilmTemplate = () => {
@@ -10,28 +9,6 @@ const createSortFilmTemplate = () => {
     </ul>`
   );
 };
-
-// export default class SortFilmComponent {
-//   constructor() {
-//     this._element = null;
-//   }
-
-//   getTemplate() {
-//     return createSortFilmTemplate();
-//   }
-
-//   getElement() {
-//     if (!this._element) {
-//       this._element = createElement(this.getTemplate());
-//     }
-
-//     return this._element;
-//   }
-
-//   removeElement() {
-//     this._element = null;
-//   }
-// }
 
 export default class MainNavigation extends AbstractComponent {
   getTemplate() {
