+++ src/const.js
@@ -14,5 +14,5 @@ export const MONTHS = [
 ];
 
 export const CARD_COUNT = 5;
-export const MOVIE_COUNT = 8;
+export const MOVIE_COUNT = 9;
 export const CARD_COUNT_TOP = 2;
