+++ src/utils/utils.js
@@ -1,26 +1,3 @@
-// export const RenderPosition = {
-//   AFTERBEGIN: `afterbegin`,
-//   BEFOREEND: `beforeend`
-// };
-// // фун-ия создания DOM узда
-// export const createElement = (template) => {
-//   const newElement = document.createElement(`div`);
-//   newElement.innerHTML = template;
-//   return newElement.firstChild;
-// };
-
-// // фун-ия вставки элемента
-// export const render = (container, element, place) => {
-//   switch (place) {
-//     case RenderPosition.AFTERBEGIN:
-//       container.prepend(element);
-//       break;
-//     case RenderPosition.BEFOREEND:
-//       container.append(element);
-//       break;
-//   }
-// };
-
 // фун-ия возвращающая случайное число [0..number)
 export const getRandomNumber = (number) => {
   return Math.floor(Math.random() * number);
